Revert "Plan 9 from Bell Labs Fourth Edition 2007-03-04"
This reverts commit 2ddf2468114832cf7379e5290d3756505dc39637.

+++ sys/src/cmd/ql/l.h
@@ -50,7 +50,7 @@ struct	Prog
 	short	line;
 	short	mark;
 	short	optab;		/* could be uchar */
-	ushort	as;
+	uchar	as;
 	char	reg;
 };
 struct	Sym
@@ -75,7 +75,7 @@ struct	Autom
 };
 struct	Optab
 {
-	ushort	as;
+	uchar	as;
 	char	a1;
 	char	a2;
 	char	a3;
